+++ apps/site-app/src/containers/MainRouter/index.tsx
@@ -4,7 +4,7 @@ import { Switch, Route } from 'react-router-dom';
 import { Accounts } from 'containers/Accounts';
 import { AddAccount } from '@thecointech/site-base/containers/AddAccount';
 import { Congratulations } from '@thecointech/site-base/containers/AddAccount/Congratulations';
-import { GAuth } from '@thecointech/site-base/containers/AddAccount/Storage/StoreOnline/Google/gauth';
+import { GAuth } from '@thecointech/site-base/containers/AddAccount/Storage/GDrive/gauth';
 
 
 export const MainRouter = (props: { location: Location }) => {
