+++ apps/site-app/src/containers/Settings/StorageOptions/index.tsx
@@ -3,7 +3,7 @@ import { Download } from '@thecointech/site-base/containers/AddAccount/Storage/S
 import { Container, Header } from "semantic-ui-react"
 import { isWallet } from '@thecointech/utilities/SignerIdent';
 import { Props as MessageProps, MaybeMessage } from "@thecointech/site-base/components/MaybeMessage"
-import { StoreGoogle, UploadState } from '@thecointech/site-base/containers/AddAccount/Storage/StoreOnline/Google';
+import { StoreGoogle, UploadState } from '@thecointech/site-base/containers/AddAccount/Storage/GDrive';
 import { useActiveAccount } from '@thecointech/shared/containers/AccountMap';
 import { FormattedMessage } from 'react-intl';
 import styles from './styles.module.less';
