+++ libs/site-base/src/containers/AddAccount/Storage/GDrive/Store.stories.tsx
@@ -1,0 +1,10 @@
+import { GDriveStore } from './Store';
+import { withAccounts } from '@thecointech/storybookutils';
+
+export default {
+  title: "Base/Storage/GDrive/Store",
+  component: GDriveStore,
+  decorators: [withAccounts()]
+};
+
+export const Store = {}
