+++ libs/site-base/src/containers/AddAccount/Storage/styles.module.less
@@ -9,10 +9,10 @@
 }
 
 .choice {
+  display: inline-block;
+  // limit the size to prevent text from expanding container
+  max-width: 171px;
   img{
-      max-width: 150px;
-      min-width: 100px;
-      width: 100%;
       margin: auto;
       display: block;
   }
